refactor : split app bar components into separate files
