+++ client/src/features/app-bar/components/notificationsMenu.styles.js
@@ -1,0 +1,12 @@
+const styles = {
+  notificationBox: {
+    display: "flex",
+    flexDirection: "column",
+    width: "100%",
+  },
+  dismissButton: {
+    mt: 1,
+  },
+};
+
+export default styles;
