+++ client/src/features/app-bar/components/profileMenu.styles.js
@@ -1,0 +1,7 @@
+const styles = {
+  menuIcon: {
+    mr: 1,
+  },
+};
+
+export default styles;
