+++ client/src/features/app-bar/components/tafSelector.styles.js
@@ -1,0 +1,12 @@
+const styles = {
+  TAFSelectorBox: {
+    display: "flex",
+    justifyContent: "center",
+    flexGrow: 1,
+  },
+  formControl: {
+    minWidth: 300,
+  },
+};
+
+export default styles;
